Artack Type Design
Artack is a clean, contemporary, and artistic sans-serif typeface including regular weights with a thin, uniform line width and details that enhance legibility. Artack is designed to be used as the brand typeface of Virtual Museum of Broken Relationships. It serves for installation art, signage and wall lettering inside the museum, as well as for visual poetry, posters, web banners, packaging, and more.